Remove foreign key constraint fk_ole_ds_item_t_ole_ds_holdings_t_holdings_id
This constraint prevented the ole_ds_item_t table (and dependent tables) from being built if for some reason the holdings record was not included in ole_ds_holdings_t. One case that caused that was when an instance, holdings, and item record were created after the LDP update began, but before the holdings and item tables were extracted. This addresses Bugzilla ticket 26817.